+++ src/polyrhythm.rs
@@ -263,7 +263,7 @@ impl Polyrhythm {
             .into_iter()
             .map(|tone| {
                 let interval = Interval::from_semitones(tone.offset)?;
-                Ok(base_pitch.transpose(&interval))
+                base_pitch.transpose(&interval)
             })
             .collect()
     }
@@ -331,7 +331,7 @@ impl Polyrhythm {
             .into_iter()
             .map(|offset| {
                 let interval = Interval::from_semitones(offset)?;
-                Ok(base_pitch.transpose(&interval))
+                base_pitch.transpose(&interval)
             })
             .collect();